statistics for caliber heuristic
Looks like the caliber heuristic really brings a marginal performance
improvement (on the harder test instance I am using).

Heap operations now take 49.7% of the total runtime and calculating the
calibers 1.6%.

Without the caliber heuristic heap operations took 52.7% of the total
runtime.

Even though the empirical analysis I've done is as lousy as it gets, it
does not look like implementing this is worth the effort.  Having a look
at what kind of performance impact a multi-level bucket instead of a
heap would have, is still an interesting topic.